api/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic[email]==2.7.1  # Added for email validation
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
redis==5.0.2
celery==5.3.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
httpx==0.27.0
